+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Import subscribers feature is now maintained by additional User Import add-on
+- Import button only available when User Import add-on is installed
+
 ## [2.0.0-alpha1] - 2023-01-31
 ### Added
 - [#50] Implemented detailed error text in send-log